refactor (Applying CGLib Proxy and JDK Proxy): move class location
* 클래스 위치 수정
