+++ src/test/java/next/hello/HelloCGLibProxyTest.java
@@ -1,6 +1,7 @@
 package next.hello;
 
 import next.interceptor.UppercaseMethodInterceptor;
+import next.wrapper.EnhancerWrapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
