+++ src/test/java/next/hello/JdkProxyTest.java
@@ -2,6 +2,7 @@ package next.hello;
 
 import next.handler.UpperCaseInvocationHandler;
 import next.interceptor.UppercaseMethodInterceptor;
+import next.wrapper.EnhancerWrapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
